adding Nirav's classifier to the current revision of the code
